refactor(frontend): fold presenter and panel view-state passthroughs
